DE/requirements.txt
chromedriver
selenium
beautifulsoup4
requests
pandas
numpy
python-jobspy
tqdm
google-generativeai